feat(login): add working functions to login / register
